requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
matplotlib>=3.3.0
tqdm>=4.50.0
Pillow>=8.0.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
pandas>=1.2.0
imageio>=2.9.0